Philbert WPF update ter
Maj ajoutrecette
Maj xml

+++ Bdd_Rezmer_Revol/WpfBdd/FenGestionCooking.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -12,8 +13,9 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using MySql.Data.MySqlClient;
-using System.ComponentModel;
 using System.Data;
+using System.Xml;
+using System.Xml.Serialization;
 
 
 namespace WpfBdd
@@ -25,6 +27,7 @@ namespace WpfBdd
     {
         MySqlConnection connexion;
         DataTable tableTop5;
+        DataTable tableCommande;
         static string recetteSupprimee;
         static string cuisinierSupprime;
         public FenGestionCooking(MySqlConnection connexion)
@@ -110,20 +113,70 @@ namespace WpfBdd
 
         private void btnSupprimerRecette_Click(object sender, RoutedEventArgs e)
         {
-            MySqlCommand commande = this.connexion.CreateCommand();
-            commande.CommandText = "DELETE FROM recette WHERE nomRecette=\"" + recetteSupprimee + "\";"; ;
-            commande.ExecuteNonQuery();
-            MessageBox.Show("La recette " + recetteSupprimee + " a bien été supprimée.", "Information", MessageBoxButton.OK, MessageBoxImage.Asterisk);
-            comboBoxRecette.Items.Remove(comboBoxRecette.SelectedItem);
+            if (comboBoxRecette.SelectedItem != null)
+            {
+                MySqlCommand commande = this.connexion.CreateCommand();
+                commande.CommandText = "DELETE FROM recette WHERE nomRecette=\"" + recetteSupprimee + "\";"; ;
+                commande.ExecuteNonQuery();
+                MessageBox.Show("La recette " + recetteSupprimee + " a bien été supprimée.", "Information", MessageBoxButton.OK, MessageBoxImage.Asterisk);
+                comboBoxRecette.Items.Remove(comboBoxRecette.SelectedItem);
+                comboBoxRecette.SelectedItem = null;
+                commande.CommandText = "SELECT nomRecette, type, prixDeVente, compteur, idCuisinier FROM recette ORDER BY compteur DESC;";
+                commande.ExecuteNonQuery();
+                tableTop5 = new DataTable("Top 5 des recettes");
+                MySqlDataAdapter dataAdp = new MySqlDataAdapter(commande);
+                dataAdp.Fill(tableTop5);
+                dataGridTop5.ItemsSource = tableTop5.DefaultView;
+            }
+            else
+            {
+                MessageBox.Show("Vous n'avez rien sélectionné", "Attention", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+            
         }
 
         private void btnSupprimerCuisinier_Click(object sender, RoutedEventArgs e)
         {
+            if (comboBoxCuisinier.SelectedItem != null)
+            {
+                MySqlCommand commande = this.connexion.CreateCommand();
+                commande.CommandText = "DELETE FROM cuisinier WHERE idCuisinier=\"" + cuisinierSupprime + "\";"; ;
+                commande.ExecuteNonQuery();
+                MessageBox.Show("Le cusinier " + cuisinierSupprime + " a bien été supprimé.", "Information", MessageBoxButton.OK, MessageBoxImage.Asterisk);
+                comboBoxCuisinier.Items.Remove(comboBoxCuisinier.SelectedItem);
+                comboBoxCuisinier.SelectedItem = null;
+                commande.CommandText = "SELECT nomRecette, type, prixDeVente, compteur, idCuisinier FROM recette ORDER BY compteur DESC;";
+                commande.ExecuteNonQuery();
+                tableTop5 = new DataTable("Top 5 des recettes");
+                MySqlDataAdapter dataAdp = new MySqlDataAdapter(commande);
+                dataAdp.Fill(tableTop5);
+                dataGridTop5.ItemsSource = tableTop5.DefaultView;
+            }
+            else
+            {
+                MessageBox.Show("Vous n'avez rien sélectionné", "Attention", MessageBoxButton.OK, MessageBoxImage.Warning);
+            }
+        }
+
+        private void btnXML_Click(object sender, RoutedEventArgs e)
+        {
+            string path = "D:/ESILV/2019 2020/Semestre 6/Base de données/TD/DM";
+            XmlWriter writer = XmlWriter.Create(path+"/test.xml");
+            writer.Close();
             MySqlCommand commande = this.connexion.CreateCommand();
-            commande.CommandText = "DELETE FROM cuisinier WHERE idCuisinier=\"" + cuisinierSupprime + "\";"; ;
+            commande.CommandText = "SELECT produit.idProduit, nomProduit, produit.refFournisseur, nomF FROM produit, fournisseur WHERE produit.refFournisseur = fournisseur.refFournisseur AND produit.stockActuel < produit.stockMin ORDER BY produit.refFournisseur;";
             commande.ExecuteNonQuery();
-            MessageBox.Show("Le cusinier " + cuisinierSupprime + " a bien été supprimé.", "Information", MessageBoxButton.OK, MessageBoxImage.Asterisk);
-            comboBoxCuisinier.Items.Remove(comboBoxCuisinier.SelectedItem);
+            tableCommande = new DataTable("Fournisseur");
+            MySqlDataAdapter dataAdp = new MySqlDataAdapter(commande);
+            dataAdp.Fill(tableCommande);
+            tableCommande.WriteXml(path+"/test.xml");
+            commande.CommandText = "SELECT produit.idProduit, nomProduit, produit.refFournisseur, nomF FROM produit, fournisseur WHERE produit.refFournisseur = fournisseur.refFournisseur AND produit.stockActuel < produit.stockMin ORDER BY produit.idProduit;";
+            commande.ExecuteNonQuery();
+            tableCommande = new DataTable("Produit");
+            dataAdp = new MySqlDataAdapter(commande);
+            dataAdp.Fill(tableCommande);
+            tableCommande.WriteXml("D:/ESILV/2019 2020/Semestre 6/Base de données/TD/DM/test.xml");
+            MessageBox.Show("Le fichier .XML a bien été généré dans le répertoire " + path,"Information");
         }
     }
 }
